+++ pkg/backend/display/display.go
@@ -57,6 +57,22 @@ func printPermalink(out io.Writer, opts Options, message, permalink string) {
 	}
 }
 
+// stampEvents stamps each event with a sequence number and a timestamp, this is to ensure consistent
+// timestamps between the events written to event log files and the events displayed in the terminal.
+func stampEvents(events <-chan engine.Event) <-chan engine.StampedEvent {
+	stampedEvents := make(chan engine.StampedEvent)
+	go func() {
+		var sequence int
+		for e := range events {
+			timestamp := int(time.Now().Unix())
+			stampedEvents <- engine.StampedEvent{Event: e, Sequence: sequence, Timestamp: timestamp}
+			sequence++
+		}
+		close(stampedEvents)
+	}()
+	return stampedEvents
+}
+
 // ShowEvents reads events from the `events` channel until it is closed, displaying each event as
 // it comes in. Once all events have been read from the channel and displayed, it closes the `done`
 // channel so the caller can await all the events being written.
@@ -64,22 +80,44 @@ func ShowEvents(
 	op string, action apitype.UpdateKind, stack tokens.StackName, proj tokens.PackageName,
 	permalink string, events <-chan engine.Event, done chan<- bool, opts Options, isPreview bool,
 ) {
+	// As we show events we need to stamp them with a sequence number and a timestamp, this is so we get
+	// consistent timestamps written in both startEventLogger and ShowJSONEvents. If we feed these two
+	// separate functions the raw `engine.Event` channel to do their own timestamping we can end up with
+	// timestamp differences.
+	stampedEvents := stampEvents(events)
+
 	if opts.EventLogPath != "" {
-		events, done = startEventLogger(events, done, opts)
+		stampedEvents, done = startEventLogger(stampedEvents, done, opts)
 	}
 
 	// Need to filter the engine events here to exclude any internal events.
-	events = channel.FilterRead(events, func(e engine.Event) bool {
-		return !e.Internal()
+	stampedEvents = channel.FilterRead(stampedEvents, func(e engine.StampedEvent) bool {
+		return !e.Event.Internal()
 	})
 
 	streamPreview := cmdutil.IsTruthy(os.Getenv("PULUMI_ENABLE_STREAMING_JSON_PREVIEW"))
 
+	// If we're in non-preview JSON display mode we need to show the stamped events, for anything else we need
+	// to show the raw events. We work out here if we're transforming the stamped events back to raw events so
+	// that we can consistently use `rawEvents` in both `ShowPreviewDigest` and the non-json display modes. We
+	// can't always create rawEvents because if we're in ShowJSONEvents we'll have two consumers competing for
+	// the stamped events.
+	var rawEvents chan engine.Event
+	if !opts.JSONDisplay || (isPreview && !streamPreview) {
+		rawEvents = make(chan engine.Event)
+		go func() {
+			for e := range stampedEvents {
+				rawEvents <- e.Event
+			}
+			close(rawEvents)
+		}()
+	}
+
 	if opts.JSONDisplay {
 		if isPreview && !streamPreview {
-			ShowPreviewDigest(events, done, opts)
+			ShowPreviewDigest(rawEvents, done, opts)
 		} else {
-			ShowJSONEvents(events, done, opts)
+			ShowJSONEvents(stampedEvents, done, opts)
 		}
 		return
 	}
@@ -90,27 +128,24 @@ func ShowEvents(
 
 	switch opts.Type {
 	case DisplayDiff:
-		ShowDiffEvents(op, events, done, opts)
+		ShowDiffEvents(op, rawEvents, done, opts)
 	case DisplayProgress:
-		ShowProgressEvents(op, action, stack, proj, permalink, events, done, opts, isPreview)
-	case DisplayQuery:
-		contract.Failf("DisplayQuery can only be used in query mode, which should be invoked " +
-			"directly instead of through ShowEvents")
+		ShowProgressEvents(op, action, stack, proj, permalink, rawEvents, done, opts, isPreview)
 	case DisplayWatch:
-		ShowWatchEvents(op, events, done, opts)
+		ShowWatchEvents(op, rawEvents, done, opts)
 	default:
 		contract.Failf("Unknown display type %d", opts.Type)
 	}
 }
 
-func logJSONEvent(encoder *json.Encoder, event engine.Event, opts Options, seq int) error {
-	apiEvent, err := ConvertEngineEvent(event, false /* showSecrets */)
+func logJSONEvent(encoder *json.Encoder, event engine.StampedEvent, opts Options) error {
+	apiEvent, err := ConvertEngineEvent(event.Event, false /* showSecrets */)
 	if err != nil {
 		return err
 	}
 
-	apiEvent.Sequence = seq
-	apiEvent.Timestamp = int(time.Now().Unix())
+	apiEvent.Sequence = event.Sequence
+	apiEvent.Timestamp = event.Timestamp
 	// If opts.Color == "never" (i.e. NO_COLOR is specified or --color=never), clean up the color directives
 	// from the emitted events.
 	if opts.Color == colors.Never {
@@ -131,7 +166,9 @@ func logJSONEvent(encoder *json.Encoder, event engine.Event, opts Options, seq i
 	return encoder.Encode(apiEvent)
 }
 
-func startEventLogger(events <-chan engine.Event, done chan<- bool, opts Options) (<-chan engine.Event, chan<- bool) {
+func startEventLogger(
+	events <-chan engine.StampedEvent, done chan<- bool, opts Options,
+) (<-chan engine.StampedEvent, chan<- bool) {
 	// Before moving further, attempt to open the log file.
 	//
 	// Try setting O_APPEND to see if that helps with the malformed reads we've been seeing in automation api:
@@ -142,21 +179,19 @@ func startEventLogger(events <-chan engine.Event, done chan<- bool, opts Options
 		return events, done
 	}
 
-	outEvents, outDone := make(chan engine.Event), make(chan bool)
+	outEvents, outDone := make(chan engine.StampedEvent), make(chan bool)
 	go func() {
 		defer close(done)
 		defer func() {
 			contract.IgnoreError(logFile.Close())
 		}()
 
-		sequence := 0
 		encoder := json.NewEncoder(logFile)
 		encoder.SetEscapeHTML(false)
 		for e := range events {
-			if err = logJSONEvent(encoder, e, opts, sequence); err != nil {
+			if err = logJSONEvent(encoder, e, opts); err != nil {
 				logging.V(7).Infof("failed to log event: %v", err)
 			}
-			sequence++
 
 			outEvents <- e