+++ sdk/go/pulumi-language-go/language_test.go
@@ -21,6 +21,7 @@ import (
 	"io"
 	"os"
 	"os/exec"
+	"path/filepath"
 	"sync"
 	"testing"
 
@@ -176,15 +177,23 @@ var expectedFailures = map[string]string{
 	"l1-builtin-can": "pulumi#18570 Support can in Go program generation",
 
 	// pulumi/pulumi#18345
-	"l1-keyword-overlap":                  "outputs are not cast correctly from pcl to their pulumi types",                                           //nolint:lll
-	"l2-plain":                            "cannot use &plain.DataArgs{…} (value of type *plain.DataArgs) as plain.DataArgs value in struct literal", //nolint:lll
-	"l2-map-keys":                         "cannot use &plain.DataArgs{…} (value of type *plain.DataArgs) as plain.DataArgs value in struct literal", //nolint:lll
-	"l2-resource-parent-inheritance":      "not implemented yet",
+	"l1-keyword-overlap":                  "outputs are not cast correctly from pcl to their pulumi types",                                                 //nolint:lll
+	"l2-plain":                            "cannot use &plain.DataArgs{…} (value of type *plain.DataArgs) as plain.DataArgs value in struct literal",       //nolint:lll
+	"l2-map-keys":                         "cannot use &plain.DataArgs{…} (value of type *plain.DataArgs) as plain.DataArgs value in struct literal",       //nolint:lll
 	"l2-component-program-resource-ref":   "pulumi#18140: cannot use ref.Value (variable of type pulumi.StringOutput) as string value in return statement", //nolint:lll
 	"l2-component-component-resource-ref": "pulumi#18140: cannot use ref.Value (variable of type pulumi.StringOutput) as string value in return statement", //nolint:lll
 	"l2-component-call-simple":            "pulumi#18202: syntax error: unexpected / in parameter list; possibly missing comma or )",                       //nolint:lll
-	"l2-component-property-deps":          "pulumi#18202: syntax error: unexpected / in parameter list; possibly missing comma or )",                       //nolint:lll                                                   //nolint:lll
 	"l2-resource-invoke-dynamic-function": "pulumi#18423: pulumi.Interface{} unexpected {, expected )",                                                     //nolint:lll
+}
+
+// Add program overrides here for programs that can't yet be generated correctly due to programgen bugs.
+var programOverrides = map[string]*testingrpc.PrepareLanguageTestsRequest_ProgramOverride{
+	// TODO[pulumi/pulumi#18202]: Delete this override when the programgen issue is addressed.
+	"l2-component-property-deps": {
+		Paths: []string{
+			filepath.Join("testdata", "overrides", "l2-component-property-deps"),
+		},
+	},
 }
 
 func TestLanguage(t *testing.T) {
@@ -228,6 +237,7 @@ func TestLanguage(t *testing.T) {
 				Replacement: "/ROOT/artifacts",
 			},
 		},
+		ProgramOverrides: programOverrides,
 	})
 	require.NoError(t, err)